+++ Packages/com.walletconnect.core/CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Dispose WalletConnect on `ApplicationQuit`. This will also dispose the instance when exiting play mode in the Editor.
+
+### Changed
+
+- Don’t pause WebSocket when app loses focus
+
+### Fixed
+
+- Device type detection on Android
+
 ## [3.1.1] - 2024-03-25
 
 ### Added